Telegram/requirements.txt
python-dotenv==1.0.1
python-telegram-bot==21.0.0
sqlalchemy==2.0.23
requests==2.31.0
